requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncio==3.4.3
certifi==2023.11.17
colorama==0.4.6
distro==1.8.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.24.4
openai==1.3.7
pandas==2.0.3
Pillow==10.1.0
pydantic==2.5.2
pydantic-core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
retrying==1.3.4
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
